requirements.txt
apispec==5.1.0; python_version >= "3.6"
certifi==2021.5.30; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
charset-normalizer==2.0.6; python_full_version >= "3.6.0" and python_version >= "3"
falcon-apispec==0.3.0
falcon==3.0.1; python_version >= "3.5"
gunicorn==20.1.0; python_version >= "3.5"
idna==3.2; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
jinja2==3.0.1; python_version >= "3.6"
markupsafe==2.0.1; python_version >= "3.6"
numpy==1.21.1; python_version >= "3.7" and python_full_version >= "3.7.1"
pandas-datapackage-reader==0.18.0
pandas==1.3.3; python_full_version >= "3.7.1"
pyproj==3.2.1; python_version >= "3.7"
python-dateutil==2.8.2; python_full_version >= "3.7.1"
pytz==2021.1; python_full_version >= "3.7.1"
pyyaml==5.4.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
requests==2.26.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
six==1.16.0; python_full_version >= "3.7.1"
swagger-ui-py==0.4.1
urllib3==1.22; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
psycopg2==2.9.1
